+++ DatabaseUsers/src/main/java/com/grupo01/spring/service/UsuarioService.java
@@ -8,7 +8,7 @@ public interface UsuarioService {
 
 	public Usuario crearUsuario(Usuario usuario) throws Exception;
 	
-	public Optional<Usuario> usuarioById(long user_id);
+	public Optional<Usuario> usuarioById(int user_id);
 	
     public Optional<Usuario> usuarioByMail(String mail);
